Release v0.1.0 (#86)
* Issue #14: Add semantic version

This commit addresses #14 and adds semantic versioning to our bot,
but it also does some other important things:

- Create a CHANGELOG.md
- Rewire Dockerfile for faster builds. Specifically move
requirements installation to the last possible line
- Remove unnecessary requirements from requirements.txt

* Address Issues #60, #63, #65

This commit accomplishes the following things:
- It address some typos spotted. It's -> its and editting ->
editing in the appropriate places
- It uses the testing's bot's telegram token for local
development
- It changes up the formatting of the display of the status
message by opening up the design more

* Addresses #59 and #69

This commit addresses the issues listed above. Specifically:

- Variable name is fixed (refraction wasn't done fully) so that
status's user's display name is accessed
- Remove all old mention styles and use the implemented function
to get mentions (refraction wasn't done fully)
- Fix command usage description in invitations for a notify group

* Address #70

This commit adds the ability for the bot to delete the /notify
command from the creator of the notify group if the bot has
sufficient permissions to delete messages in a group chat.
If it doesn't it just silently fails

* Addresses #82

- This commit outputs the db user password which is a random
password
- It also fixes up the prompt for the user when there are no
status users to display the status of
- And it also adds a requirement that is should have been there
for pymongo[srv]

* Addresses #77

This commit allows multiple invites to a notify groups instead of
just one. This commit also:

- Removes mentions of Issue #31 and completes an incomplete
refraction
- Adds more user prompts to give the user more feedback on button
clicks.

However, the invite_to_notify_group handlers still need a fair bit
more work and issues to improve the feature

* Adds an /about command

* Add release information to changelog