Release 1.0.0 (#100)
* Issue #14: Add semantic version

This commit addresses #14 and adds semantic versioning to our bot,
but it also does some other important things:

- Create a CHANGELOG.md
- Rewire Dockerfile for faster builds. Specifically move
requirements installation to the last possible line
- Remove unnecessary requirements from requirements.txt

* Address Issues #60, #63, #65

This commit accomplishes the following things:
- It address some typos spotted. It's -> its and editting ->
editing in the appropriate places
- It uses the testing's bot's telegram token for local
development
- It changes up the formatting of the display of the status
message by opening up the design more

* Addresses #59 and #69

This commit addresses the issues listed above. Specifically:

- Variable name is fixed (refraction wasn't done fully) so that
status's user's display name is accessed
- Remove all old mention styles and use the implemented function
to get mentions (refraction wasn't done fully)
- Fix command usage description in invitations for a notify group

* Address #70

This commit adds the ability for the bot to delete the /notify
command from the creator of the notify group if the bot has
sufficient permissions to delete messages in a group chat.
If it doesn't it just silently fails

* Addresses #82

- This commit outputs the db user password which is a random
password
- It also fixes up the prompt for the user when there are no
status users to display the status of
- And it also adds a requirement that is should have been there
for pymongo[srv]

* Addresses #77

This commit allows multiple invites to a notify groups instead of
just one. This commit also:

- Removes mentions of Issue #31 and completes an incomplete
refraction
- Adds more user prompts to give the user more feedback on button
clicks.

However, the invite_to_notify_group handlers still need a fair bit
more work and issues to improve the feature

* Adds an /about command

* Fixes the v0.1.0 link for CHANGELOG.md

* Address #51 (#88)

This commit ensures better UX flow
for conversation handlers, ensuring
that other command entries in other
chats don't interrupt a current
conversation running in a different
chat

* Address #64 (#89)

This issue adds further guides for
users who were unable to run a
command because they either did not
have the bot started or the bot was
blocked, so the bot could not send a
private message to the user.

* Address #74 (#90)

This commit allows the user to list
specific notify groups instead of
listing all the notify groups in a group
chat. However, this does NOT add
this information to the help command.
We will add that later.

* Addresses #76 (#91)

This commit sorts the notify groups
alphabetically if no notify groups are
specified. It also adds the hint that
the notify groups are sorted to the
message from the bot

* Addresses #76 (#92)

This commit addresses the issue of
bolding the title of notify groups when
they are listed. These bold titles are
to replace the pre-formatted titles.

* Addresses #68 (#95)

Add a button that can be used in
order to use the alternative notify
group name that the bot suggests.

* Addresses #94 (#96)

This commit adds emojis to some buttons that didn't have one AND it also fixes a bug where messages weren't being sent due to broken inline keyboards

* Addresses #4 (#97)

This commit changes the help message into a help menu, allowing for more detailed information on what each command does.

* Update CHANGELOG.md (#98)

* Addresses #93 (#99)

This commit allows lowercase arguments to be entered for commands.

* Update changelog and semantic versioning to reflect a new release

* Upgrade python version, PSNAWP, and init APIWrapper for prod only

* Fix a typo in the /about command text